Added MobX to project and created stores
