+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"E:\\React-apps\\crypto-converter\\src\\index.tsx":"1","E:\\React-apps\\crypto-converter\\src\\App.tsx":"2"},{"size":221,"mtime":1606995805726,"results":"3","hashOfConfig":"4"},{"size":5241,"mtime":1607010914210,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1982r0d",{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\React-apps\\crypto-converter\\src\\index.tsx",[],"E:\\React-apps\\crypto-converter\\src\\App.tsx",[]]
+[{"E:\\React-apps\\crypto-converter\\src\\index.tsx":"1","E:\\React-apps\\crypto-converter\\src\\App.tsx":"2","E:\\React-apps\\crypto-converter\\src\\components\\index.ts":"3","E:\\React-apps\\crypto-converter\\src\\components\\CryptoTable\\index.tsx":"4","E:\\React-apps\\crypto-converter\\src\\components\\ConverterBlock\\index.tsx":"5","E:\\React-apps\\crypto-converter\\src\\styles.ts":"6","E:\\React-apps\\crypto-converter\\src\\stores\\index.ts":"7","E:\\React-apps\\crypto-converter\\src\\stores\\converterStore.ts":"8","E:\\React-apps\\crypto-converter\\src\\stores\\currenciesStore.ts":"9"},{"size":337,"mtime":1607034242419,"results":"10","hashOfConfig":"11"},{"size":635,"mtime":1607040010876,"results":"12","hashOfConfig":"11"},{"size":118,"mtime":1607023442952,"results":"13","hashOfConfig":"11"},{"size":2110,"mtime":1607040014056,"results":"14","hashOfConfig":"11"},{"size":1698,"mtime":1607038091648,"results":"15","hashOfConfig":"11"},{"size":731,"mtime":1607027524901,"results":"16","hashOfConfig":"11"},{"size":225,"mtime":1607034123355,"results":"17","hashOfConfig":"11"},{"size":64,"mtime":1607033965023,"results":"18","hashOfConfig":"11"},{"size":1037,"mtime":1607039977988,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"1982r0d",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\React-apps\\crypto-converter\\src\\index.tsx",[],["39","40"],"E:\\React-apps\\crypto-converter\\src\\App.tsx",[],"E:\\React-apps\\crypto-converter\\src\\components\\index.ts",[],"E:\\React-apps\\crypto-converter\\src\\components\\CryptoTable\\index.tsx",[],"E:\\React-apps\\crypto-converter\\src\\components\\ConverterBlock\\index.tsx",[],"E:\\React-apps\\crypto-converter\\src\\styles.ts",[],"E:\\React-apps\\crypto-converter\\src\\stores\\index.ts",[],"E:\\React-apps\\crypto-converter\\src\\stores\\converterStore.ts",[],"E:\\React-apps\\crypto-converter\\src\\stores\\currenciesStore.ts",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"no-global-assign","no-unsafe-negation"]
