+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as CryptoTable } from './CryptoTable';
+export { default as ConverterBlock } from './ConverterBlock';
