+++ src/stores/converterStore.ts
@@ -1,0 +1,7 @@
+
+
+class ConverterStore {
+
+}
+
+export default ConverterStore
