+++ src/types/index.ts
@@ -1,0 +1,7 @@
+export type TCoin = {
+  name: string;
+  fullName: string;
+  ImageUrl: string;
+  price: number;
+  valume24hour: number;
+};
