+++ tsconfig.json
@@ -18,7 +18,8 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx"
+		"jsx": "react-jsx",
+		"experimentalDecorators": true
   },
   "include": [
     "src"
